+++ src/main/java/com/example/springmvcdemo/dev/repository/BookAuthorRepository.java
@@ -3,6 +3,6 @@ package com.example.springmvcdemo.dev.repository;
 import com.example.springmvcdemo.dev.model.BookAuthor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookAuthorRepository extends JpaRepository<BookAuthor, Long> {
+public interface BookAuthorRepository extends JpaRepository<BookAuthor, Integer> {
 
 }
